Introduce `ActiveModel::Enum`
This moves most of the implementation of ActiveRecord::Enum to
ActiveModel::Enum. This excludes:
* scopes, as scopes aren't supported by ActiveModel.
* method conflict detection, as this is difficult to implement in
  ActiveModel.

The current implementation supports using integers for values and
strings for labels, if defined. This required moving
ActiveRecord::AttributeMethods::BeforeTypeCast to ActiveModel as well.

The tests were copied from ActiveRecord and modified for ActiveModel.
Scoping and method conflict detection tests were excluded.